requirements.txt
aiofiles
aiohttp
discord.py
python-dotenv
exa-py
Flask
google-generativeai
--find-links https://storage.googleapis.com/generativeai-downloads/wheels/
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pypdf
psutil
requests
serpapi
sympy
tavily-python
